serbot/pyproject.toml
[tool.poetry]
name = "serbot"
version = "1.0.0"
description = "service bot"
authors = ["Sevent Tella <7onetella@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
slackclient = "^2.9.3"
slackeventsapi = "^2.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
